__temp/chatwoot_demo_generator/requirements.txt
faker>=19.0.0
requests>=2.31.0
python-dotenv>=1.0.0
